+++ src/revlanguage/monitors/Mntr_StochasticCharacterMapping.cpp
@@ -55,7 +55,7 @@ void Mntr_StochasticCharacterMapping::constructInternalObject( void )
     bool               is             = static_cast<const RlBoolean &>( include_simmap->getRevObject() ).getValue();
     bool               sd             = static_cast<const RlBoolean &>( use_simmap_default->getRevObject() ).getValue();
     const std::string& sep            = static_cast<const RlString  &>( separator->getRevObject()      ).getValue();
-    int                print_gen      = static_cast<const Natural   &>( printgen->getRevObject()       ).getValue();
+    int                print_gen      = (int)static_cast<const Natural   &>( printgen->getRevObject()       ).getValue();
     bool               app            = static_cast<const RlBoolean &>( append->getRevObject()         ).getValue();
